project.tsx 다크 모드 기능 추가 및 파일명 변경(Project.tsx -> project.tsx), pages폴더로 이동
